add force ചെറ്റത്തരം  subscription

+++ plugins/forceSubscribe.py
@@ -15,17 +15,26 @@ def _onUnMuteRequest(client, cb):
   chat_id = cb.message.chat.id
   chat_db = sql.fs_settings(chat_id)
   if chat_db:
-    channel = chat_db.channel
+    channel_s = chat_db.channel.split(".")
     chat_member = client.get_chat_member(chat_id, user_id)
     if chat_member.restricted_by:
       if chat_member.restricted_by.id == (client.get_me()).id:
+        n_subscription_s = []
+        for channel in channel_s:
           try:
             client.get_chat_member(channel, user_id)
+          except UserNotParticipant:
+            n_subscription_s.append(channel)
+          if len(n_subscription_s) > 0:
+            client.answer_callback_query(
+              cb.id,
+              text="❗ Join the mentioned 'channel(s)' and press the 'UnMute Me' button again.",
+              show_alert=True
+            )
+          else:
             client.unban_chat_member(chat_id, user_id)
             if cb.message.reply_to_message.from_user.id == user_id:
               cb.message.delete()
-          except UserNotParticipant:
-            client.answer_callback_query(cb.id, text="❗ Join the mentioned 'channel' and press the 'UnMute Me' button again.", show_alert=True)
       else:
         client.answer_callback_query(cb.id, text="❗ You are muted by admins for other reasons.", show_alert=True)
     else:
@@ -44,13 +53,23 @@ def _check_member(client, message):
   if chat_db:
     user_id = message.from_user.id
     if not client.get_chat_member(chat_id, user_id).status in ("administrator", "creator") and not user_id in Config.SUDO_USERS:
-      channel = chat_db.channel
-      try:
-        client.get_chat_member(channel, user_id)
-      except UserNotParticipant:
+      channels = chat_db.channel.split(".")
+      n_subscription_s = []
+      for channel in channels:
+        try:
+          client.get_chat_member(channel, user_id)
+        except UserNotParticipant:
+          n_subscription_s.append(channel)
+        except ChatAdminRequired:
+          client.send_message(chat_id, text=f"❗ **I am not an admin in @{channel}**\n__Make me admin in the channel and add me again.\n#Leaving this chat...__")
+          sql.disapprove(chat_id)
+          client.leave_chat(chat_id)
+          return
+      if len(n_subscription_s) > 0:
+        mmo = "@" + " @".join(n_subscription_s)
         try:
           sent_message = message.reply_text(
-              "{}, you are **not subscribed** to my [channel](https://t.me/{}) yet. Please [join](https://t.me/{}) and **press the button below** to unmute yourself.".format(message.from_user.mention, channel, channel),
+              "{}, you are **not subscribed** to my channels: {mmo} yet. Please [join](https://t.me/{}) and **press the button below** to unmute yourself.".format(message.from_user.mention, channel, channel),
               disable_web_page_preview=True,
               reply_markup=InlineKeyboardMarkup(
                   [[InlineKeyboardButton("UnMute Me", callback_data="onUnMuteRequest")]]
@@ -60,9 +79,6 @@ def _check_member(client, message):
         except ChatAdminRequired:
           sent_message.edit("❗ **I am not an admin here.**\n__Make me admin with ban user permission and add me again.\n#Leaving this chat...__")
           client.leave_chat(chat_id)
-      except ChatAdminRequired:
-        client.send_message(chat_id, text=f"❗ **I am not an admin in @{channel}**\n__Make me admin in the channel and add me again.\n#Leaving this chat...__")
-        client.leave_chat(chat_id)
 
 
 @Client.on_message(filters.command(["forcesubscribe", "fsub"]) & ~filters.private)
